requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
python-dateutil==2.8.2
Pillow==10.1.0
